refactor: reorganize surface factory imports and update module structure
