+++ optiland/surfaces/factories/__init__.py
@@ -1,0 +1,3 @@
+# flake8: noqa
+
+from .surface_factory import SurfaceFactory
